@@ -3,29 +3,27 @@
   "description": "ESLint Shareable Config",
   "version": "22.0.0",
   "author": "Takahito Nakano",
+  "dependencies": {
+    "globals": "^15.14.0"
+  },
   "devDependencies": {
+    "@vitest/eslint-plugin": "^1.1.25",
     "cross-env": "^7.0.3",
-    "eslint": "^8.57.1",
+    "eslint": "^9.20.0",
     "eslint-config-prettier": "^10.0.1",
     "eslint-find-rules": "^5.0.0",
-    "eslint-plugin-jest": "^28.11.0",
     "husky": "^9.1.7",
-    "jest": "^29.7.0",
     "lint-staged": "^15.4.3",
     "npm-run-all2": "^7.0.2",
-    "prettier": "^3.4.2"
+    "prettier": "^3.4.2",
+    "vitest": "^3.0.5"
   },
   "files": [
     "base.js",
     "browser.js",
     "index.js",
-    "node.js",
-    "rules"
+    "node.js"
   ],
-  "jest": {
-    "testEnvironment": "node",
-    "verbose": true
-  },
   "keywords": [
     "eslint",
     "eslint-config"
@@ -33,7 +31,7 @@
   "license": "MIT",
   "main": "index.js",
   "peerDependencies": {
-    "eslint": "^8.57.1"
+    "eslint": "^9.20.0"
   },
   "repository": {
     "type": "git",
@@ -51,6 +49,6 @@
     "lint:eslint": "eslint . --report-unused-disable-directives",
     "lint:prettier": "prettier . --check",
     "prepare": "node .husky/install.mjs",
-    "test": "cross-env ESLINT_CONFIG_PRETTIER_NO_DEPRECATED=true jest"
+    "test": "cross-env ESLINT_CONFIG_PRETTIER_NO_DEPRECATED=true vitest"
   }
 }
