@@ -1,15 +1,17 @@
 'use strict';
 
+const globals = require('globals');
+const base = require('./base');
+
+/** @type {import("eslint").Linter.Config} */
 module.exports = {
-  extends: ['./base'].map(require.resolve),
-  env: {
-    browser: false,
-    node: true,
-  },
-  parserOptions: {
-    sourceType: 'script',
-    ecmaFeatures: {
-      impliedStrict: false,
+  ...base,
+  languageOptions: {
+    ...base.languageOptions,
+    sourceType: 'commonjs',
+    globals: {
+      ...base.languageOptions?.globals,
+      ...globals.node,
     },
   },
 };
